API Integration Tests

+++ src/controllers/projects.js
@@ -1,21 +1,29 @@
 const ProjectModel = require('../models/projects')
+const mongoose = require('mongoose')
 
 async function get(req, res) {
+  if (mongoose.connection.readyState !== 1) {
+    return res.status(503).json({ message: 'Error connecting to the database' })
+  }
   const { id } = req.params
 
   const obj = id ? { _id: id } : null
 
   try {
     const projects = await ProjectModel.find(obj)
-
     console.log(projects)
-    res.send(projects)
+    await res.send(projects)
+
+    res.status(200)
   } catch (error) {
     res.status(500).json({ message: 'Error fetching projects', error })
   }
 }
 
 async function post(req, res) {
+  if (mongoose.connection.readyState !== 1) {
+    return res.status(503).json({ message: 'Error connecting to the database' })
+  }
   try {
     const { name, url, description, image } = req.body
 
@@ -26,7 +34,7 @@ async function post(req, res) {
       description,
     })
 
-    register.save()
+    await register.save()
 
     const id = register._id
 
@@ -35,20 +43,33 @@ async function post(req, res) {
       id: id,
     })
   } catch (error) {
-    res.status(500).json({ message: 'Error creating project', error })
+    if (error.name === 'ValidationError') {
+      res.status(400).json({ message: 'Validation failed' })
+    } else {
+      res.status(500).json({ message: 'Error creating project', error })
+    }
   }
 }
 
 async function put(req, res) {
-  const { id } = req.params
-  const { name, url, description, image } = req.body
-
+  if (mongoose.connection.readyState !== 1) {
+    return res.status(503).json({ message: 'Error connecting to the database' })
+  }
   try {
-    const project = await ProjectModel.findById(id)
+    const { id } = req.params
+    const update = req.body
 
-    if (!project) {
+    console.log('update no controller: ', update)
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
       return res.status(404).json({ message: 'Project not found' })
     }
+    if (!update || Object.keys(update).length == 0) {
+      return res.status(400).json({ message: 'Nothing to update' })
+    }
+
+    const { name, url, description, image } = req.body
+    const project = await ProjectModel.findById({ _id: id })
 
     project.name = name !== undefined ? name : project.name
     project.url = url !== undefined ? url : project.url
@@ -58,7 +79,7 @@ async function put(req, res) {
 
     const updatedProject = await project.save()
 
-    res.status(200).json({
+    res.status(201).json({
       message: 'Success. Document updated',
       data: updatedProject,
     })
@@ -72,10 +93,14 @@ async function put(req, res) {
 }
 
 async function remove(req, res) {
+  if (mongoose.connection.readyState !== 1) {
+    return res.status(503).json({ message: 'Error connecting to the database' })
+  }
   try {
     const { id } = req.params
-    console.log('id  do controller delete: ', id)
-
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).json({ message: 'Project not found' })
+    }
     await ProjectModel.deleteOne({ _id: id })
     res.status(204).send()
   } catch (error) {
